requirements.txt
ctransformers==0.2.27
chromadb==0.4.22
langchain
tiktoken
sentence_transformers==2.2.2
InstructorEmbedding==1.0.1
unstructured==0.5.6
python-dotenv
openai
pytesseract
fitz
pillow
deep_translator
flask
langchain-google-genai
langchain_openai
watchdog==3.0.0
